backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-cloud-vision==3.4.4
google-generativeai==0.3.1
requests==2.31.0